+++ lib/src/controller/schedule_notificaitons.dart
@@ -21,6 +21,7 @@
 ///      <uses-permission android:name="android.permission.VIBRATE" />
 ///      <uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED"/>
 
+import 'dart:typed_data';
 import 'dart:async' show Future;
 import 'dart:math' show Random;
 import 'dart:typed_data' show Int64List;
@@ -58,6 +59,8 @@ export 'package:flutter_local_notifications/flutter_local_notifications.dart'
         InboxStyleInformation,
         IOSNotificationDetails,
         IOSNotificationAttachment,
+        MacOSNotificationDetails,
+        MacOSNotificationAttachment,
         MediaStyleInformation,
         NotificationAppLaunchDetails,
         NotificationDetails,
@@ -98,9 +101,11 @@ class ScheduleNotifications with HandleError {
     bool? autoCancel,
     bool? ongoing,
     Color? color,
-    AndroidBitmap? largeIcon,
+    AndroidBitmap<Object>? largeIcon,
     bool? onlyAlertOnce,
     bool? showWhen,
+    int? when,
+    bool? usesChronometer,
     bool? channelShowBadge,
     bool? showProgress,
     int? maxProgress,
@@ -116,6 +121,10 @@ class ScheduleNotifications with HandleError {
     int? timeoutAfter,
     String? category,
     bool? fullScreenIntent,
+    String? shortcutId,
+    Int32List? additionalFlags,
+    String? subText,
+    String? tag,
     RepeatInterval? repeatInterval,
     Day? day,
     SelectNotificationCallback? onSelectNotification,
@@ -132,6 +141,8 @@ class ScheduleNotifications with HandleError {
     String? soundFile,
     int? badgeNumber,
     List<IOSNotificationAttachment>? attachments,
+    String? subtitle,
+    String? threadIdentifier,
   }) {
     if (appIcon == null || appIcon.trim().isEmpty) {
       // Assign the app's icon.
@@ -162,6 +173,8 @@ class ScheduleNotifications with HandleError {
     _largeIcon = largeIcon;
     _onlyAlertOnce = onlyAlertOnce ?? false;
     _showWhen = showWhen ?? true;
+//    _when =
+    _usesChronometer = usesChronometer ?? false;
     _channelShowBadge = channelShowBadge ?? true;
     _showProgress = showProgress ?? false;
     _maxProgress = maxProgress ?? 0;
@@ -178,6 +191,10 @@ class ScheduleNotifications with HandleError {
     _timeoutAfter = timeoutAfter;
     _category = category;
     _fullScreenIntent = fullScreenIntent ?? false;
+    _shortcutId = shortcutId;
+    _additionalFlags = additionalFlags;
+    _subText = subText;
+    _tag = tag;
     _repeatInterval = repeatInterval;
     _day = day;
     _selectNotificationCallback = onSelectNotification;
@@ -194,6 +211,8 @@ class ScheduleNotifications with HandleError {
     _soundFile = soundFile;
     _badgeNumber = badgeNumber;
     _attachments = attachments;
+    _subtitle = subtitle;
+    _threadIdentifier = threadIdentifier;
   }
 
   /// The icon representing the app implementing the notifications.
@@ -230,11 +249,11 @@ class ScheduleNotifications with HandleError {
   bool? _autoCancel;
   bool? _ongoing;
   Color? _color;
-  AndroidBitmap? _largeIcon;
+  AndroidBitmap<Object>? _largeIcon;
   bool? _onlyAlertOnce;
   bool? _showWhen;
   int? _when;
-  bool? usesChronometer;
+  bool? _usesChronometer;
   bool? _channelShowBadge;
   bool? _showProgress;
   int? _maxProgress;
@@ -250,6 +269,10 @@ class ScheduleNotifications with HandleError {
   int? _timeoutAfter;
   String? _category;
   bool? _fullScreenIntent;
+  String? _shortcutId;
+  Int32List? _additionalFlags;
+  String? _subText;
+  String? _tag;
   RepeatInterval? _repeatInterval;
   Day? _day;
 //  Time _notificationTime;
@@ -267,6 +290,9 @@ class ScheduleNotifications with HandleError {
   String? _soundFile;
   int? _badgeNumber;
   List<IOSNotificationAttachment>? _attachments;
+  List<MacOSNotificationAttachment>? _macAttachments;
+  String? _subtitle;
+  String? _threadIdentifier;
 
   @mustCallSuper
   Future<bool?> init({
@@ -290,9 +316,11 @@ class ScheduleNotifications with HandleError {
     bool? autoCancel,
     bool? ongoing,
     Color? color,
-    AndroidBitmap? largeIcon,
+    AndroidBitmap<Object>? largeIcon,
     bool? onlyAlertOnce,
     bool? showWhen,
+    int? when,
+    bool? usesChronometer,
     bool? channelShowBadge,
     bool? showProgress,
     int? maxProgress,
@@ -308,6 +336,10 @@ class ScheduleNotifications with HandleError {
     int? timeoutAfter,
     String? category,
     bool? fullScreenIntent,
+    String? shortcutId,
+    Int32List? additionalFlags,
+    String? subText,
+    String? tag,
     RepeatInterval? repeatInterval,
     Day? day,
     Time? notificationTime,
@@ -325,6 +357,8 @@ class ScheduleNotifications with HandleError {
     String? soundFile,
     int? badgeNumber,
     List<IOSNotificationAttachment>? attachments,
+    String? subtitle,
+    String? threadIdentifier,
   }) async {
     // No need to continue.
     if (_init!) {
@@ -402,6 +436,12 @@ class ScheduleNotifications with HandleError {
     if (showWhen != null) {
       _showWhen = showWhen;
     }
+    if (when != null) {
+      _when = when;
+    }
+    if (usesChronometer != null) {
+      _usesChronometer = usesChronometer;
+    }
     if (channelShowBadge != null) {
       _channelShowBadge = channelShowBadge;
     }
@@ -447,14 +487,23 @@ class ScheduleNotifications with HandleError {
     if (fullScreenIntent != null) {
       _fullScreenIntent = fullScreenIntent;
     }
-    if (repeatInterval != null) {
-      _repeatInterval = repeatInterval;
+    if (shortcutId != null) {
+      _shortcutId = shortcutId;
+    }
+    if (additionalFlags != null) {
+      _additionalFlags = additionalFlags;
+    }
+    if (subText != null) {
+      _subText = subText;
+    }
+    if (tag != null) {
+      _tag = tag;
     }
     if (day != null) {
       _day = day;
     }
-    onSelectNotification ??= _selectNotificationCallback ??
-        _onSelectNotification as Future<dynamic> Function(String?)?;
+    onSelectNotification ??=
+        _selectNotificationCallback ?? _onSelectNotification;
     requestAlertPermission ??= _requestAlertPermission;
     requestSoundPermission ??= _requestSoundPermission;
     requestBadgePermission ??= _requestBadgePermission;
@@ -480,7 +529,12 @@ class ScheduleNotifications with HandleError {
     if (attachments != null) {
       _attachments = attachments;
     }
-
+    if (subtitle != null) {
+      _subtitle = subtitle;
+    }
+    if (threadIdentifier != null) {
+      _threadIdentifier = threadIdentifier;
+    }
     //
     try {
       final initializationSettingsIOS = IOSInitializationSettings(
@@ -551,6 +605,37 @@ class ScheduleNotifications with HandleError {
     return request;
   }
 
+  // Future<bool> resolvePlatformSpecificImplementation<T extends MethodChannelFlutterLocalNotificationsPlugin>( {
+  //   bool alert = true,
+  //       bool badge = true,
+  //   bool sound = true,
+  // }) async {
+  // assert(_init!, 'ScheduleNotifications: Failed to call init() first!');
+  //
+  // if (!_init!) {
+  // return false;
+  // }
+  // T? implementation;
+  // try {
+  // implementation = _flutterLocalNotificationsPlugin
+  //     ?.resolvePlatformSpecificImplementation<
+  // T>();
+  // } catch (ex) {
+  // getError(ex);
+  // }
+  // var request = implementation != null;
+  // if (request) {
+  // request = await (implementation!.requestPermissions(
+  // alert: alert,
+  // badge: badge,
+  // sound: sound,
+  // ) as Future<bool>);
+  // }
+  // return request;
+  // }
+  //
+  // }
+
   // plugin
   FlutterLocalNotificationsPlugin? _flutterLocalNotificationsPlugin;
 
@@ -578,9 +663,11 @@ class ScheduleNotifications with HandleError {
     bool? autoCancel,
     bool? ongoing,
     Color? color,
-    AndroidBitmap? largeIcon,
+    AndroidBitmap<Object>? largeIcon,
     bool? onlyAlertOnce,
     bool? showWhen,
+    int? when,
+    bool? usesChronometer,
     bool? channelShowBadge,
     bool? showProgress,
     int? maxProgress,
@@ -596,12 +683,19 @@ class ScheduleNotifications with HandleError {
     int? timeoutAfter,
     String? category,
     bool? fullScreenIntent,
+    String? shortcutId,
+    Int32List? additionalFlags,
+    String? subText,
+    String? tag,
     bool? presentAlert,
     bool? presentSound,
     bool? presentBadge,
     String? soundFile,
     int? badgeNumber,
     List<IOSNotificationAttachment>? attachments,
+    List<MacOSNotificationAttachment>? macAttachments,
+    String? subtitle,
+    String? threadIdentifier,
   }) {
     //
     final notificationSpecifics = _notificationDetails(
@@ -626,6 +720,8 @@ class ScheduleNotifications with HandleError {
       largeIcon,
       onlyAlertOnce,
       showWhen,
+      when,
+      usesChronometer,
       channelShowBadge,
       showProgress,
       maxProgress,
@@ -641,12 +737,19 @@ class ScheduleNotifications with HandleError {
       timeoutAfter,
       category,
       fullScreenIntent,
+      shortcutId,
+      additionalFlags,
+      subText,
+      tag,
       presentAlert,
       presentSound,
       presentBadge,
       soundFile,
       badgeNumber,
       attachments,
+      macAttachments,
+      subtitle,
+      threadIdentifier,
     );
 
     if (notificationSpecifics == null) {
@@ -674,6 +777,8 @@ class ScheduleNotifications with HandleError {
     return id;
   }
 
+  @Deprecated('Deprecated due to problems with time zones. Use zonedSchedule '
+      'instead.')
   int schedule(
     TZDateTime? schedule, {
     int? id,
@@ -698,9 +803,10 @@ class ScheduleNotifications with HandleError {
     bool? autoCancel,
     bool? ongoing,
     Color? color,
-    AndroidBitmap? largeIcon,
+    AndroidBitmap<Object>? largeIcon,
     bool? onlyAlertOnce,
     bool? showWhen,
+    int? when,
     bool? channelShowBadge,
     bool? showProgress,
     int? maxProgress,
@@ -716,12 +822,135 @@ class ScheduleNotifications with HandleError {
     int? timeoutAfter,
     String? category,
     bool? fullScreenIntent,
+    String? shortcutId,
+    Int32List? additionalFlags,
+    String? subText,
+    String? tag,
     bool? presentAlert,
     bool? presentSound,
     bool? presentBadge,
     String? soundFile,
     int? badgeNumber,
     List<IOSNotificationAttachment>? attachments,
+    List<MacOSNotificationAttachment>? macAttachments,
+    String? subtitle,
+    String? threadIdentifier,
+  }) =>
+      zonedSchedule(
+        schedule,
+        id: id,
+        title: title,
+        body: body,
+        payload: payload,
+        androidAllowWhileIdle: androidAllowWhileIdle,
+        uiLocalNotificationDateInterpretation:
+            uiLocalNotificationDateInterpretation,
+        matchDateTimeComponents: matchDateTimeComponents,
+        icon: icon,
+        importance: importance,
+        priority: priority,
+        styleInformation: styleInformation,
+        playSound: playSound,
+        sound: sound,
+        enableVibration: enableVibration,
+        vibrationPattern: vibrationPattern,
+        groupKey: groupKey,
+        setAsGroupSummary: setAsGroupSummary,
+        groupAlertBehavior: groupAlertBehavior,
+        autoCancel: autoCancel,
+        ongoing: ongoing,
+        color: color,
+        largeIcon: largeIcon,
+        onlyAlertOnce: onlyAlertOnce,
+        showWhen: showWhen,
+        when: when,
+        channelShowBadge: channelShowBadge,
+        showProgress: showProgress,
+        maxProgress: maxProgress,
+        progress: progress,
+        indeterminate: indeterminate,
+        channelAction: channelAction,
+        enableLights: enableLights,
+        ledColor: ledColor,
+        ledOnMs: ledOnMs,
+        ledOffMs: ledOffMs,
+        ticker: ticker,
+        visibility: visibility,
+        timeoutAfter: timeoutAfter,
+        category: category,
+        fullScreenIntent: fullScreenIntent,
+        shortcutId: shortcutId,
+        additionalFlags: additionalFlags,
+        subText: subText,
+        tag: tag,
+        presentAlert: presentAlert,
+        presentSound: presentSound,
+        presentBadge: presentBadge,
+        soundFile: soundFile,
+        badgeNumber: badgeNumber,
+        attachments: attachments,
+        macAttachments: macAttachments,
+        subtitle: subtitle,
+        threadIdentifier: threadIdentifier,
+      );
+
+  int zonedSchedule(
+    TZDateTime? schedule, {
+    int? id,
+    String? title,
+    String? body,
+    String? payload,
+    bool? androidAllowWhileIdle,
+    UILocalNotificationDateInterpretation?
+        uiLocalNotificationDateInterpretation,
+    DateTimeComponents? matchDateTimeComponents,
+    String? icon,
+    Importance? importance,
+    Priority? priority,
+    StyleInformation? styleInformation,
+    bool? playSound,
+    AndroidNotificationSound? sound,
+    bool? enableVibration,
+    List<int>? vibrationPattern,
+    String? groupKey,
+    bool? setAsGroupSummary,
+    GroupAlertBehavior? groupAlertBehavior,
+    bool? autoCancel,
+    bool? ongoing,
+    Color? color,
+    AndroidBitmap<Object>? largeIcon,
+    bool? onlyAlertOnce,
+    bool? showWhen,
+    int? when,
+    bool? usesChronometer,
+    bool? channelShowBadge,
+    bool? showProgress,
+    int? maxProgress,
+    int? progress,
+    bool? indeterminate,
+    AndroidNotificationChannelAction? channelAction,
+    bool? enableLights,
+    Color? ledColor,
+    int? ledOnMs,
+    int? ledOffMs,
+    String? ticker,
+    NotificationVisibility? visibility,
+    int? timeoutAfter,
+    String? category,
+    bool? fullScreenIntent,
+    String? shortcutId,
+    Int32List? additionalFlags,
+    String? subText,
+    String? tag,
+    bool? presentAlert,
+    bool? presentSound,
+    bool? presentBadge,
+    String? soundFile,
+    int? badgeNumber,
+    List<IOSNotificationAttachment>? attachments,
+    List<MacOSNotificationAttachment>? macAttachments,
+    String? subtitle,
+    String? threadIdentifier,
   }) {
     // May have already been supplied.
     schedule ??= _schedule as TZDateTime;
@@ -753,6 +982,8 @@ class ScheduleNotifications with HandleError {
       largeIcon,
       onlyAlertOnce,
       showWhen,
+      when,
+      usesChronometer,
       channelShowBadge,
       showProgress,
       maxProgress,
@@ -768,12 +999,19 @@ class ScheduleNotifications with HandleError {
       timeoutAfter,
       category,
       fullScreenIntent,
+      shortcutId,
+      additionalFlags,
+      subText,
+      tag,
       presentAlert,
       presentSound,
       presentBadge,
       soundFile,
       badgeNumber,
       attachments,
+      macAttachments,
+      subtitle,
+      threadIdentifier,
     );
 
     if (notificationSpecifics == null) {
@@ -828,9 +1066,11 @@ class ScheduleNotifications with HandleError {
     bool? autoCancel,
     bool? ongoing,
     Color? color,
-    AndroidBitmap? largeIcon,
+    AndroidBitmap<Object>? largeIcon,
     bool? onlyAlertOnce,
     bool? showWhen,
+    int? when,
+    bool? usesChronometer,
     bool? channelShowBadge,
     bool? showProgress,
     int? maxProgress,
@@ -846,12 +1086,19 @@ class ScheduleNotifications with HandleError {
     int? timeoutAfter,
     String? category,
     bool? fullScreenIntent,
+    String? shortcutId,
+    Int32List? additionalFlags,
+    String? subText,
+    String? tag,
     bool? presentAlert,
     bool? presentSound,
     bool? presentBadge,
     String? soundFile,
     int? badgeNumber,
     List<IOSNotificationAttachment>? attachments,
+    List<MacOSNotificationAttachment>? macAttachments,
+    String? subtitle,
+    String? threadIdentifier,
   }) {
     repeatInterval ??= _repeatInterval;
 
@@ -882,6 +1129,8 @@ class ScheduleNotifications with HandleError {
       largeIcon,
       onlyAlertOnce,
       showWhen,
+      when,
+      usesChronometer,
       channelShowBadge,
       showProgress,
       maxProgress,
@@ -897,12 +1146,19 @@ class ScheduleNotifications with HandleError {
       timeoutAfter,
       category,
       fullScreenIntent,
+      shortcutId,
+      additionalFlags,
+      subText,
+      tag,
       presentAlert,
       presentSound,
       presentBadge,
       soundFile,
       badgeNumber,
       attachments,
+      macAttachments,
+      subtitle,
+      threadIdentifier,
     );
 
     if (notificationSpecifics == null) {
@@ -922,6 +1178,7 @@ class ScheduleNotifications with HandleError {
           repeatInterval,
           notificationSpecifics,
           payload: payload,
+          androidAllowWhileIdle: androidAllowWhileIdle!,
         );
       } catch (ex) {
         id = -1;
@@ -955,9 +1212,11 @@ class ScheduleNotifications with HandleError {
     bool? autoCancel,
     bool? ongoing,
     Color? color,
-    AndroidBitmap? largeIcon,
+    AndroidBitmap<Object>? largeIcon,
     bool? onlyAlertOnce,
     bool? showWhen,
+    int? when,
+    bool? usesChronometer,
     bool? channelShowBadge,
     bool? showProgress,
     int? maxProgress,
@@ -973,12 +1232,19 @@ class ScheduleNotifications with HandleError {
     int? timeoutAfter,
     String? category,
     bool? fullScreenIntent,
+    String? shortcutId,
+    Int32List? additionalFlags,
+    String? subText,
+    String? tag,
     bool? presentAlert,
     bool? presentSound,
     bool? presentBadge,
     String? soundFile,
     int? badgeNumber,
     List<IOSNotificationAttachment>? attachments,
+    List<MacOSNotificationAttachment>? macAttachments,
+    String? subtitle,
+    String? threadIdentifier,
   }) {
     schedule ??= _schedule as TZDateTime;
 
@@ -1009,6 +1275,8 @@ class ScheduleNotifications with HandleError {
       largeIcon,
       onlyAlertOnce,
       showWhen,
+      when,
+      usesChronometer,
       channelShowBadge,
       showProgress,
       maxProgress,
@@ -1024,12 +1292,19 @@ class ScheduleNotifications with HandleError {
       timeoutAfter,
       category,
       fullScreenIntent,
+      shortcutId,
+      additionalFlags,
+      subText,
+      tag,
       presentAlert,
       presentSound,
       presentBadge,
       soundFile,
       badgeNumber,
       attachments,
+      macAttachments,
+      subtitle,
+      threadIdentifier,
     );
 
     if (notificationSpecifics == null) {
@@ -1088,9 +1363,11 @@ class ScheduleNotifications with HandleError {
     bool? autoCancel,
     bool? ongoing,
     Color? color,
-    AndroidBitmap? largeIcon,
+    AndroidBitmap<Object>? largeIcon,
     bool? onlyAlertOnce,
     bool? showWhen,
+    int? when,
+    bool? usesChronometer,
     bool? channelShowBadge,
     bool? showProgress,
     int? maxProgress,
@@ -1106,12 +1383,19 @@ class ScheduleNotifications with HandleError {
     int? timeoutAfter,
     String? category,
     bool? fullScreenIntent,
+    String? shortcutId,
+    Int32List? additionalFlags,
+    String? subText,
+    String? tag,
     bool? presentAlert,
     bool? presentSound,
     bool? presentBadge,
     String? soundFile,
     int? badgeNumber,
     List<IOSNotificationAttachment>? attachments,
+    List<MacOSNotificationAttachment>? macAttachments,
+    String? subtitle,
+    String? threadIdentifier,
   }) {
     day ??= _day;
 
@@ -1144,6 +1428,8 @@ class ScheduleNotifications with HandleError {
       largeIcon,
       onlyAlertOnce,
       showWhen,
+      when,
+      usesChronometer,
       channelShowBadge,
       showProgress,
       maxProgress,
@@ -1159,12 +1445,19 @@ class ScheduleNotifications with HandleError {
       timeoutAfter,
       category,
       fullScreenIntent,
+      shortcutId,
+      additionalFlags,
+      subText,
+      tag,
       presentAlert,
       presentSound,
       presentBadge,
       soundFile,
       badgeNumber,
       attachments,
+      macAttachments,
+      subtitle,
+      threadIdentifier,
     );
 
     if (notificationSpecifics == null) {
@@ -1216,9 +1509,11 @@ class ScheduleNotifications with HandleError {
     bool? autoCancel,
     bool? ongoing,
     Color? color,
-    AndroidBitmap? largeIcon,
+    AndroidBitmap<Object>? largeIcon,
     bool? onlyAlertOnce,
     bool? showWhen,
+    int? when,
+    bool? usesChronometer,
     bool? channelShowBadge,
     bool? showProgress,
     int? maxProgress,
@@ -1234,12 +1529,19 @@ class ScheduleNotifications with HandleError {
     int? timeoutAfter,
     String? category,
     bool? fullScreenIntent,
+    String? shortcutId,
+    Int32List? additionalFlags,
+    String? subText,
+    String? tag,
     bool? presentAlert,
     bool? presentSound,
     bool? presentBadge,
     String? soundFile,
     int? badgeNumber,
     List<IOSNotificationAttachment>? attachments,
+    List<MacOSNotificationAttachment>? macAttachments,
+    String? subtitle,
+    String? threadIdentifier,
   ) {
     //
     NotificationDetails? notificationSpecifics;
@@ -1275,6 +1577,8 @@ class ScheduleNotifications with HandleError {
     largeIcon ??= _largeIcon;
     onlyAlertOnce ??= _onlyAlertOnce;
     showWhen ??= _showWhen;
+    when ??= _when;
+    usesChronometer ??= _usesChronometer;
     channelShowBadge ??= _channelShowBadge;
     showProgress ??= _showProgress;
     maxProgress ??= _maxProgress;
@@ -1290,12 +1594,23 @@ class ScheduleNotifications with HandleError {
     timeoutAfter ??= _timeoutAfter;
     category ??= _category;
     fullScreenIntent ??= _fullScreenIntent;
+    shortcutId ??= _shortcutId;
+    additionalFlags ??= _additionalFlags;
+    subText ??= _subText;
+    tag ??= _tag;
     presentAlert ??= _presentAlert;
     presentSound ??= _presentSound;
     presentBadge ??= _presentBadge;
     soundFile ??= _soundFile;
     badgeNumber ??= _badgeNumber;
     attachments ??= _attachments;
+    macAttachments ??= _macAttachments;
+    subtitle ??= _subtitle;
+    threadIdentifier ??= _threadIdentifier;
+
+    if (sound == null && soundFile != null && soundFile.trim().isNotEmpty) {
+      sound = RawResourceAndroidNotificationSound(soundFile.trim());
+    }
 
     // Play the sound if supplied a sound.
     if (playSound == null) {
@@ -1318,6 +1633,7 @@ class ScheduleNotifications with HandleError {
 
     AndroidNotificationDetails androidSettings;
     IOSNotificationDetails iOSSettings;
+    MacOSNotificationDetails macOSSettings;
 
     try {
       androidSettings = AndroidNotificationDetails(
@@ -1341,6 +1657,8 @@ class ScheduleNotifications with HandleError {
         largeIcon: largeIcon,
         onlyAlertOnce: onlyAlertOnce!,
         showWhen: showWhen!,
+        when: when,
+        usesChronometer: usesChronometer!,
         channelShowBadge: channelShowBadge!,
         showProgress: showProgress!,
         maxProgress: maxProgress!,
@@ -1355,6 +1673,11 @@ class ScheduleNotifications with HandleError {
         visibility: visibility,
         timeoutAfter: timeoutAfter,
         category: category,
+        fullScreenIntent: fullScreenIntent!,
+        shortcutId: shortcutId,
+        additionalFlags: additionalFlags,
+        subText: subText,
+        tag: tag,
       );
     } catch (ex) {
       getError(ex);
@@ -1369,11 +1692,31 @@ class ScheduleNotifications with HandleError {
         sound: soundFile,
         badgeNumber: badgeNumber,
         attachments: attachments,
+        subtitle: subtitle,
+        threadIdentifier: threadIdentifier,
+      );
+    } catch (ex) {
+      notificationSpecifics = null;
+      getError(ex);
+      return notificationSpecifics;
+    }
+
+    try {
+      macOSSettings = MacOSNotificationDetails(
+        presentAlert: presentAlert,
+        presentSound: presentSound,
+        presentBadge: presentBadge,
+        sound: soundFile,
+        badgeNumber: badgeNumber,
+        attachments: macAttachments,
+        subtitle: subtitle,
+        threadIdentifier: threadIdentifier,
       );
 
       notificationSpecifics = NotificationDetails(
         android: androidSettings,
         iOS: iOSSettings,
+        macOS: macOSSettings,
       );
     } catch (ex) {
       notificationSpecifics = null;
@@ -1383,11 +1726,11 @@ class ScheduleNotifications with HandleError {
   }
 
   /// Cancel a specific notification.
-  Future<void> cancel(int id) async {
+  Future<void> cancel(int? id, {String? tag}) async {
     if (id == null || id < 0) {
       return;
     }
-    await _flutterLocalNotificationsPlugin!.cancel(id);
+    await _flutterLocalNotificationsPlugin!.cancel(id, tag: tag);
     return;
   }
 
@@ -1402,9 +1745,9 @@ class ScheduleNotifications with HandleError {
   Future<List<PendingNotificationRequest>> pendingNotificationRequests() =>
       _flutterLocalNotificationsPlugin!.pendingNotificationRequests();
 
-  void _onSelectNotification(String payload) {
-    if (payload != null || payload.trim().isNotEmpty) {
-      showDialog(
+  Future<dynamic> _onSelectNotification(String? payload) async {
+    if (payload != null || payload!.trim().isNotEmpty) {
+      await showDialog(
         context: App.context!,
         builder: (BuildContext context) {
           return AlertDialog(
@@ -1416,5 +1759,6 @@ class ScheduleNotifications with HandleError {
         },
       );
     }
+    return payload;
   }
 }